regalloc: fix bugs in graph coloring
Misunderstanding: wasn't clear what kind of nodes was in the
interference graph. Clearly now we know that both registers AND vars
are in the initially constructed interference graph.

BUT we only want to really perform the register allocation between
variables -- the registers are just there to saturate the variable
colors initially.